apps/web/pyproject.toml
[project]
name = "manganize-web"
version = "0.1.0"
description = "Web application for manganize"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.20",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.15.0",
    "pydantic-settings>=2.7.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    "manganize-core",
]

[tool.uv.sources]
manganize-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
